requirements.txt
Django==2.2.13
Pillow==7.1.1
django-ckeditor==5.9.0
django-mptt
django-crispy-forms==1.9.0
